Author post permission problem fixed

+++ app/Http/Controllers/Author/PostController.php
@@ -84,11 +84,19 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
-        return view('admin.post.show',compact('post'));
+        if($post->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
+        return view('author.post.show',compact('post'));
     }
 
     public function edit(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         $categories=Category::all();
         $tags=Tag::all();
         return view('author.post.edit',compact('post','categories','tags'));
@@ -96,6 +104,10 @@ class PostController extends Controller
 
     public function update(Request $request, Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         $this->validate($request,[
             'title'=>'required',
             'image'=>'required',
@@ -153,6 +165,10 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('you are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         if (Storage::disk('public')->exists('post/'.$post->image)) {
             Storage::disk('public')->delete('post/'.$post->image);
         }
